Update javadoc in CsvHandler class
